feat: add shared simulator utilities
