+++ src/genecoder/d2sim_adapter.py
@@ -8,7 +8,8 @@ from .api import Simulator
 from .simulators import register_simulator as _register_simulator
 
 from .random_utils import make_rng
-from .nanopore_sim import _simulate_adapter, _run_external
+from .nanopore_sim import _simulate_adapter
+from .simulator_utils import _run_external
 
 
 class _D2SIM:
